requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
django-ckeditor==6.5.1
django-dotenv==1.4.2
django-js-asset==2.0.0
django-modeltranslation==0.18.7
django-mptt==0.14.0
django-solo==2.0.0
easy-thumbnails==2.8.5
Pillow==9.4.0
python-dotenv==0.21.1
python-gettext==4.1
sqlparse==0.4.3
typing_extensions==4.4.0
